+++ MyWebApi/MyWebApi/Repositories/SystemUserRepository.cs
@@ -111,7 +111,7 @@ namespace WebApi.Repositories
 
                 invitor.InvitedUsersCount++;
 
-                var bonus = invitor.HasPremium ? 0.05 : 0;
+                var bonus = invitor.HasPremium ? 0.05f : 0f;
                 var multiplier = 1;
 
                 if (invitor.InvitedUsersCount > 10)
@@ -126,20 +126,20 @@ namespace WebApi.Repositories
                 else if (invitor.InvitedUsersCount == 3 || invitor.InvitedUsersCount % 3 == 0)
                 {
                     if (multiplier == 1)
-                        invitor.InvitedUsersBonus = 0.15 + bonus;
+                        invitor.InvitedUsersBonus = 0.15f + bonus;
                     await TopUpPointBalance(invitor.Id, 1199 * multiplier, $"User {invitor.Id} has invited % 3 users");
                 }
                 else if (invitor.InvitedUsersCount == 7 || invitor.InvitedUsersCount % 7 == 0)
                 {
                     if (multiplier == 1)
-                        invitor.InvitedUsersBonus = 0.35 + bonus;
+                        invitor.InvitedUsersBonus = 0.35f + bonus;
                     await TopUpPointBalance(invitor.Id, 1499 * multiplier, $"User {invitor.Id} has invited % 7 users");
                 }
                 else if (invitor.InvitedUsersCount == 10 || invitor.InvitedUsersCount % 10 == 0)
                 {
                     if (multiplier == 1)
                     {
-                        invitor.InvitedUsersBonus = 0.5 + bonus;
+                        invitor.InvitedUsersBonus = 0.5f + bonus;
                         // 1499 will then turn into 1999 due to premium purchase reward
                         await TopUpPointBalance(invitor.Id, 1499, $"User {invitor.Id} has invited % 10 users");
                         //Adds + 10 random effects to users inventory
@@ -155,7 +155,7 @@ namespace WebApi.Repositories
                     await TopUpPointBalance(invitor.Id, (int)(200 + (200 * bonus) * multiplier), $"User {model.Id} was invited by user {invitor.Id}");
                 }
 
-                user.BonusIndex = 1.5;
+                user.BonusIndex = 1.5f;
                 user.ParentId = invitor.Id;
 
                 _contx.Users.Update(user);
@@ -1353,7 +1353,7 @@ namespace WebApi.Repositories
                 .FirstOrDefaultAsync();
         }
 
-        public async Task<int> TopUpPointBalance(long userId, int points, string description = "")
+        public async Task<float> TopUpPointBalance(long userId, float points, string description = "")
         {
             var time = DateTime.SpecifyKind(DateTime.UtcNow, DateTimeKind.Utc);
             var userBalance = await GetUserWalletBalance(userId);
@@ -1396,7 +1396,7 @@ namespace WebApi.Repositories
             return userBalance.Points;
         }
 
-        private async Task CreateUserBalance(long userId, int points, DateTime time)
+        private async Task CreateUserBalance(long userId, float points, DateTime time)
         {
            var userBalance = new Balance(userId, points, time);
 
@@ -1441,22 +1441,22 @@ namespace WebApi.Repositories
             return userBalance.PersonalityPoints;
         }
 
-        private async Task<bool> RegisterUserPurchaseInPoints(long userId, int points, string description)
+        private async Task<bool> RegisterUserPurchaseInPoints(long userId, float points, string description)
         {
             return await RegisterUserPurchase(userId, points, description, Currency.Points);
         }
 
-        private async Task<bool> RegisterUserPurchaseInPP(long userId, int points, string description)
+        private async Task<bool> RegisterUserPurchaseInPP(long userId, float points, string description)
         {
             return await RegisterUserPurchase(userId, points, description, Currency.OceanPoints);
         }
 
-        private async Task<bool> RegisterPurchaseInRealMoney(long userId, int points, string description, Currency currency)
+        private async Task<bool> RegisterPurchaseInRealMoney(long userId, float points, string description, Currency currency)
         {
             return await RegisterUserPurchase(userId, points, description, currency);
         }
 
-        private async Task<bool> RegisterUserPurchase(long userId, int amount, string description, Currency currency)
+        private async Task<bool> RegisterUserPurchase(long userId, float amount, string description, Currency currency)
         {
             var purchase = new Transaction
             {
@@ -1515,7 +1515,7 @@ namespace WebApi.Repositories
                 return DateTime.MinValue;
         }
 
-        public async Task<DateTime> GrantPremiumToUser(long userId, int cost, int dayDuration, Currency currency)
+        public async Task<DateTime> GrantPremiumToUser(long userId, float cost, int dayDuration, Currency currency)
         {
             var timeNow = DateTime.SpecifyKind(DateTime.UtcNow, DateTimeKind.Utc);
             var premiumFutureExpirationDate = DateTime.SpecifyKind(DateTime.UtcNow.AddDays(dayDuration), DateTimeKind.Utc);
@@ -3674,7 +3674,7 @@ namespace WebApi.Repositories
             await _contx.SaveChangesAsync();
         }
 
-        public async Task<bool> PurchaseEffectAsync(long userId, int effectId, int points, Currency currency, short count=1)
+        public async Task<bool> PurchaseEffectAsync(long userId, int effectId, float points, Currency currency, short count=1)
         {
             var balance = await GetUserWalletBalance(userId);
